fix: Changed request operand conversion to accept float values

+++ api/views.py
@@ -13,14 +13,14 @@ import math
 @api_view(['POST'])
 def add(request):
   if request.method == 'POST':
-    operands = [int(num) for num in request.data["operands"]]
+    operands = [float(num) for num in request.data["operands"]]
     result = sum(operands)
     return Response(data=result, status=status.HTTP_200_OK)
 
 @api_view(['POST'])
 def divide(request):
   if request.method == 'POST':
-    operands = [int(num) for num in request.data["operands"]]
+    operands = [float(num) for num in request.data["operands"]]
     result = operands[0]
     for i in range(1, len(operands)):
       result = result / operands[i]
@@ -29,7 +29,7 @@ def divide(request):
 @api_view(['POST'])
 def multiply(request):
   if request.method == 'POST':
-    operands = [int(num) for num in request.data["operands"]]
+    operands = [float(num) for num in request.data["operands"]]
     result = operands[0]
     for i in range(1, len(operands)):
       result = result * operands[i]
@@ -44,14 +44,14 @@ def random_string(request):
 @api_view(['POST'])
 def square_root(request):
   if request.method == 'POST':
-    operand = int(request.data["operands"][0])
+    operand = float(request.data["operands"][0])
     result = math.sqrt(operand)
     return Response(data=result, status=status.HTTP_200_OK)
 
 @api_view(['POST'])
 def subtract(request):
   if request.method == 'POST':
-    operands = [int(num) for num in request.data["operands"]]
+    operands = [float(num) for num in request.data["operands"]]
     result = operands[0]
     for i in range(1, len(operands)):
       result = result - operands[i]
